+++ backend/libs/shared-types/src/assist.contracts.js
@@ -1,0 +1,12 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.ASSIST_PATTERNS = void 0;
+exports.ASSIST_PATTERNS = {
+    SESSION_START: 'assist.session.start',
+    SESSION_END: 'assist.session.end',
+    SESSION_GET: 'assist.session.get',
+    INTERVIEW_CREATE: 'assist.interview.create',
+    INTERVIEW_LIST: 'assist.interview.list',
+    INTERVIEW_GET: 'assist.interview.get',
+};
+//# sourceMappingURL=assist.contracts.js.map